requirements.txt
langchain
openai
PyPDF2
faiss-cpu
streamlit
python-dotenv
tiktoken